Introduce GameEngineStatus::WaitingForEvent
Add this GameEngineStatus which replaces GameEventType. If
RuleResult::WaitingForEvent has to specify which event type it is
waiting for, then our engine has to know about all the different event
types we might need. Additionally, we might need more than one event
type which the current enum does not support.