Introduce CompositeRule
This struct allows us to compose rules together. The rules are applied
to the game state in order, waiting for events as necessary. Once a
rule completes, it moves on to the next rule in the chain. If a rule
is waiting for an event, it will not proceed to the next rule until
the event is consumed.